Make separate start_utils helper to declutter the start module
