+++ alita/plugins/initial.py
@@ -16,7 +16,6 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 
-
 from pyrogram import filters
 from pyrogram.errors import RPCError
 from pyrogram.types import Message
